% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function COVID Limits.R
\name{Covid_limits}
\alias{Covid_limits}
\title{Title}
\usage{
Covid_limits(dataset)
}
\arguments{
\item{dataset}{a data.frame or tibble with columns "date" and "new_events"; use this function one population at a time.}
}
\value{

}
\description{
Title
}
