% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid_limits.R
\name{Covid_limits2}
\alias{Covid_limits2}
\title{Title}
\usage{
Covid_limits2(dataset)
}
\arguments{
\item{dataset}{a data.frame or tibble with columns "date" and "new_events"; use this function one population at a time.}
}
\value{

}
\description{
Title
}
