% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function COVID Phase.R
\name{Covid_Phase}
\alias{Covid_Phase}
\title{Title}
\usage{
Covid_Phase(
  PlaceRR,
  place,
  date,
  events,
  new_events,
  MIDLINE_N,
  LOWER_N,
  UPPER_N,
  Days_N,
  Day_PC,
  Date_PC,
  EPOCH,
  Phase_N,
  Phase_Count,
  MIDLINEa,
  LOWERa,
  UPPERa,
  MIDLINEb,
  LOWERb,
  UPPERb,
  Phase_Ch
)
}
\arguments{
\item{Phase_Ch}{}
}
\value{

}
\description{
Title
}
