% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function I Chart II.R
\name{I_Chart}
\alias{I_Chart}
\title{I Chart}
\usage{
I_Chart(
  date = IChart_Data$date,
  Dot = IChart_Data$Dot,
  IChart_Data = data_frame(date, Dot)
)
}
\description{
I Chart
}
\author{
Gareth Parry
}
