% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_p_chart.R
\name{plot_p_chart}
\alias{plot_p_chart}
\title{Title}
\usage{
plot_p_chart(
  data,
  title = "",
  mode1 = "lines",
  sizes = c(1, 100),
  yname = "Positivity rate (\%)"
)
}
\arguments{
\item{yname}{}
}
\value{

}
\description{
Title
}
