% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_p_chart.R
\name{postprocess_p_chart}
\alias{postprocess_p_chart}
\title{Postprocess p and p-prime chart data}
\usage{
postprocess_p_chart(
  data,
  digits = 2,
  date_label = "{date} ({weekdays(date)})\\n",
  events = "positive tests",
  trials = "tests",
  proportion = "Positivity rate",
  multiplier = 100
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Postprocess p and p-prime chart data
}
