% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_u_chart.R
\name{postprocess_u_chart}
\alias{postprocess_u_chart}
\title{Postprocess u and u-prime chart data}
\usage{
postprocess_u_chart(
  data,
  digits = 2,
  events = "close contacts",
  trials = "primary cases",
  proportion = "Rate",
  multiplier = 1,
  suffix = paste(events, "per", trials)
)
}
\arguments{
\item{proportion}{}
}
\value{

}
\description{
Postprocess u and u-prime chart data
}
