% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_p_chart.R
\name{summarize_p_chart}
\alias{summarize_p_chart}
\title{Title}
\usage{
summarize_p_chart(
  data,
  place_vars = "place",
  reference_date = max(data$date),
  digits = 4
)
}
\arguments{
\item{reference_date}{}
}
\description{
Title
}
