% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function U Prime Chart.R
\name{U_Prime}
\alias{U_Prime}
\title{Title}
\usage{
U_Prime(dataset)
}
\arguments{
\item{dataset}{a data frame containing:

    `date`: Date formatted field containing the date on which events are reported

   `n`:    The numerator for the event proportion

    `N`:    The denominator for the event proportion}
}
\value{

}
\description{
Title
}
\author{
Gareth Parry
}
