% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function U Prime Chart Funnel.R
\name{U_Prime_Funnel}
\alias{U_Prime_Funnel}
\title{Title}
\usage{
U_Prime_Funnel(
  Uchart_Data,
  to_remove = c("MR", "Sigma_i", "Zi", "Screened_MR")
)
}
\arguments{
\item{Uchart_Data}{}
}
\value{

}
\description{
Title
}
