% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_states.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{California data}
\format{
TODO: complete this
}
\usage{
us_states
}
\description{
A dataset containing California COVID deaths data
}
\keyword{datasets}
