% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function I Chart II.R
\name{I_Chart}
\alias{I_Chart}
\title{Title}
\usage{
I_Chart(
  date = IChart_Data$date,
  Dot = IChart_Data$Dot,
  IChart_Data = data_frame(date, Dot)
)
}
\arguments{
\item{IChart_Data}{}
}
\description{
Title
}
