% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_chart.R
\name{plot_run_chart}
\alias{plot_run_chart}
\title{Title}
\usage{
plot_run_chart(
  data,
  title = "",
  mode1 = "lines",
  sizes = c(1, 100),
  yname = "Positivity rate (\%)",
  xname = NA,
  legend.y = -0.25,
  multiplier = 100,
  point_name = "Observed \%",
  hoverinfo = "x+text",
  ylab.position = 0
)
}
\arguments{
\item{hoverinfo}{}
}
\value{

}
\description{
Title
}
\author{
Doug Ezra Morrison <demorrison@ucdavis.edu>
}
