% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_chart.R
\name{plot_run_chart}
\alias{plot_run_chart}
\title{Title}
\usage{
plot_run_chart(
  data,
  title = "",
  mode1 = "lines",
  yname = "Positivity rate (\%)",
  xname = NA,
  xlim = range(data$date),
  legend.y = -0.2,
  multiplier = 100,
  marker_size = ~N,
  marker_size_range = c(1, 100),
  sizemode = "area",
  point_name = "Observed \%",
  hoverinfo = "x+text",
  ylab.position = 0,
  line_width = 2,
  limit_width = line_width,
  midline_width = 1,
  download_format = "svg"
)
}
\arguments{
\item{download_format}{}
}
\value{

}
\description{
Title
}
\author{
Doug Ezra Morrison <demorrison@ucdavis.edu>
}
